[FIX] : pull dev into Feature/DataAPI/kkomin 후 에러 수정
