+++ app/src/main/java/com/example/teamtube/Adapter/SearchFragmentAdapter.kt
@@ -1,16 +1,14 @@
-package com.example.teamtube
+package com.example.teamtube.Adapter
 
 import android.content.Context
 import android.util.Log
 import androidx.recyclerview.widget.RecyclerView
-import com.example.teamtube.SearchFragmentData.SearchItem
 import com.example.teamtube.databinding.FragmentSearchItemBinding
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.constraintlayout.widget.ConstraintLayout
 import com.bumptech.glide.Glide
 import com.example.teamtube.SearchFragmentData.SearchData
