+++ app/src/main/res/layout/fragment_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".Home.Fragment.HomeFragment">
+    tools:context=".Fragment.HomeFragment">
 
     <EditText
         android:id="@+id/sf_et_search"
